+++ src/serializer.ts
@@ -1,7 +1,8 @@
 'use strict';
 import { WebviewPanel, WebviewPanelSerializer, ExtensionContext, Uri } from 'vscode';
-import CsvPreview from './csvPreview';
-import ExcelPreview from './excelPreview';
+import CsvDocumentView from './csvDocumentView';
+import ExcelDocumentView from './excelDocumentView';
+import LegacyDocumentView from './legacyDocumentView';
 
 export class CsvSerializer implements WebviewPanelSerializer {
     
@@ -12,7 +13,7 @@ export class CsvSerializer implements WebviewPanelSerializer {
     }
 
     public async deserializeWebviewPanel(webviewPanel: WebviewPanel, state: any) {
-        CsvPreview.revive(this._context, Uri.parse(state.uri), webviewPanel);
+        CsvDocumentView.revive(this._context, Uri.parse(state.uri), webviewPanel);
     }
 }
 
@@ -25,6 +26,19 @@ export class ExcelSerializer implements WebviewPanelSerializer {
     }
 
     public async deserializeWebviewPanel(webviewPanel: WebviewPanel, state: any) {
-        ExcelPreview.revive(this._context, Uri.parse(state.uri), webviewPanel);
+        ExcelDocumentView.revive(this._context, Uri.parse(state.uri), webviewPanel);
+    }
+}
+
+export class LegacySerializer implements WebviewPanelSerializer {
+    
+    private _context: ExtensionContext;
+
+    constructor(context: ExtensionContext) {
+        this._context = context;
+    }
+
+    public async deserializeWebviewPanel(webviewPanel: WebviewPanel, state: any) {
+        LegacyDocumentView.revive(this._context, Uri.parse(state.uri), webviewPanel);
     }
 }