+++ webpack.config.js
@@ -4,15 +4,15 @@
  *--------------------------------------------------------------------------------------------*/
 
 //@ts-check
-
 'use strict';
 
 const path = require('path');
+const webpack = require('webpack');
 
 /**@type {import('webpack').Configuration}*/
-const config = {
+const nodeConfig = {
     target: 'node', // vscode extensions run in a Node.js-context 📖 -> https://webpack.js.org/configuration/node/
-
+    mode: 'none',
     entry: './src/extension.ts', // the entry point of this extension, 📖 -> https://webpack.js.org/configuration/entry-context/
     output: { // the bundle is stored in the 'dist' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
         path: path.resolve(__dirname, 'dist'),
@@ -28,7 +28,7 @@ const config = {
         extensions: ['.ts', '.js']
     },
     stats: {
-        warningsFilter: warning => RegExp("node_modules/express/lib/view.js").test(warning)
+        warningsFilter: warning => RegExp("node_modules/express/lib/view.js").test(warning.message)
     },
     module: {
         rules: [{
@@ -41,4 +41,54 @@ const config = {
     },
 }
 
-module.exports = config;
+/**@type {import('webpack').Configuration}*/
+const webConfig = {
+    target: 'webworker', // web extensions run in a webworker context
+    mode: 'none',
+    entry: {
+        'extension': './src/extension.ts' // the entry point of this extension, 📖 -> https://webpack.js.org/configuration/entry-context/
+    },
+    output: { // the bundle is stored in the 'dist' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'extension.js',
+        libraryTarget: "commonjs2",
+        devtoolModuleFilenameTemplate: "../[resource-path]",
+    },
+    devtool: 'source-map',
+    externals: {
+        vscode: "commonjs vscode" // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
+    },
+    resolve: { // support reading TypeScript and JavaScript files, 📖 -> https://github.com/TypeStrong/ts-loader
+        mainFields: ['browser', 'module', 'main'],
+        extensions: ['.ts', '.js'],
+        alias: {
+			// provides alternate implementation for node module and source files
+		},
+		fallback: {
+			// Webpack 5 no longer polyfills Node.js core modules automatically.
+			// see https://webpack.js.org/configuration/resolve/#resolvefallback
+			// for the list of Node.js core module polyfills.
+            assert: require.resolve('assert'),
+            path: require.resolve('path-browserify'),
+		}
+    },
+    stats: {
+        warningsFilter: warning => RegExp("node_modules/express/lib/view.js").test(warning.message)
+    },
+	plugins: [
+	    new webpack.ProvidePlugin({
+		    process: 'process/browser', // provide a shim for the global `process` variable
+	}),
+	],
+    module: {
+        rules: [{
+            test: /\.ts$/,
+            exclude: /node_modules/,
+            use: [{
+                loader: 'ts-loader'
+            }]
+        }]
+    },
+}
+
+module.exports = [nodeConfig, webConfig];